Master DEV

+++ app/api/admin/actions/route.js
@@ -1,10 +1,8 @@
 import { NextResponse } from 'next/server';
 import { getServerSession } from 'next-auth';
 import { connectDB } from '@/lib/db/connect';
-import ExamRoomModel from '@/lib/db/models/ExamRoom.model';
-import { getQuizById } from '@/lib/moodle/get-quiz';
+import { ExamRoomModel } from '@/lib/db/models';
 import { EXAMROOM_ACTIONS } from '@/lib/tools/constants/actions';
-import { deployCloudRun, deleteCloudRun } from '@/lib/cloud/run/controls';
 import {
   createGCRJob,
   deleteGCRJob,
@@ -12,7 +10,7 @@ import {
   fixErrorJob,
 } from '@/lib/moodle/jobs';
 import { authOptions } from '@/lib/auth/options';
-import { STATUS_CODE } from '@/lib/moodle/status';
+import { STATUS_CODE, canActions } from '@/lib/moodle/state/status';
 import { updateSchedule } from '@/lib/tools/schedule';
 
 export async function POST(req) {
@@ -38,6 +36,8 @@ export async function POST(req) {
         // start container
         try {
           const nroom = await ExamRoomModel.findById(room.id);
+          if (!canActions(nroom.status)) continue;
+
           if (!nroom.serviceUrl) {
             await createGCRJob(nroom.containerName);
           }
@@ -50,6 +50,8 @@ export async function POST(req) {
       for (const room of rooms) {
         try {
           const nroom = await ExamRoomModel.findById(room.id);
+          if (!canActions(nroom.status)) continue;
+
           if (nroom.serviceUrl) {
             await stopGCRJob(nroom.containerName);
           }
@@ -68,8 +70,12 @@ export async function POST(req) {
         // delete and save data
         try {
           const nroom = await ExamRoomModel.findById(room.id);
+          if (!canActions(nroom.status)) continue;
+
           if (nroom.serviceUrl) {
             await deleteGCRJob(nroom.containerName);
+          } else {
+            await deleteGCRJob(nroom.containerName, { saveData: false });
           }
         } catch (err) {
           console.error(`[ADMIN_ACTION]@@ ${err.message}`);
@@ -81,9 +87,9 @@ export async function POST(req) {
         // delete data
         try {
           const nroom = await ExamRoomModel.findById(room.id);
-          if (nroom.serviceUrl) {
-            await deleteGCRJob(nroom.containerName, { saveData: false });
-          }
+          if (!canActions(nroom.status)) continue;
+
+          await deleteGCRJob(nroom.containerName, { saveData: false });
         } catch (err) {
           console.error(`[ADMIN_ACTION]@@ ${err.message}`);
         }
@@ -94,6 +100,7 @@ export async function POST(req) {
         // try fix
         try {
           const nroom = await ExamRoomModel.findById(room.id);
+
           await fixErrorJob(nroom.containerName);
         } catch (err) {
           console.error(`[ADMIN_ACTION]@@ ${err.message}`);
@@ -105,6 +112,9 @@ export async function POST(req) {
         // reset schedule
         try {
           const nroom = await ExamRoomModel.findById(room.id);
+          if (nroom.serviceUrl) continue;
+          if (!canActions(nroom.status)) continue;
+
           nroom.status = STATUS_CODE.REGISTERED;
           await nroom.save();
           await updateSchedule(nroom.containerName);
